dataset-services/requirements.txt
config
elasticsearch
findspark
kafka-python
numpy
pandas
python-dateutil==2.8.0
urllib3==1.24.1
pyspark